Add DatabaseSource source and database_* script for oal

+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/DatabaseService.java
@@ -18,7 +18,7 @@ public class DatabaseService extends Source {
         return String.valueOf(id);
     }
 
-    @Getter @Setter private int id;
+    @Getter @Setter private long id;
     @Getter @Setter private String name;
     @Getter @Setter private String endpointName;
     @Getter @Setter private String statement;
